Update endpoint is mostly complete, need to properly update changes in deck name

+++ app/Http/Controllers/DeckController.php
@@ -8,6 +8,7 @@ use App\Card_deck;
 use Illuminate\Http\Request;
 use Auth;
 use JavaScript;
+use Carbon\Carbon;
 
 class DeckController extends Controller
 {
@@ -54,7 +55,7 @@ class DeckController extends Controller
        
         $data = $request->all();
         
-        $cards = json_decode($data['userDeck'],true);
+        $cards = $this->filterDeck($request->userDeck);
         $deck = Deck::create([
                 'user_id' => $data['user_id'],
                 'name' => $data['name'],
@@ -65,22 +66,29 @@ class DeckController extends Controller
             'card_id' => $data['lead_id'],
             'quantity' => 1
             ]);
-        foreach ($cards as $card) {
-            if ($card['selected']) {
-            $entry = Card_deck::create([
-                'deck_id' => $deck['id'],
-                'card_id' => $card['id'],
-                'quantity' => $card['quantity']
-                ]);
-        }
-        };
         
-        // JavaScript::put([
-        //     'cardlist' => $deck['cards']
-            
-        //     ]);
+        $this->createCardDecks ($cards,$deck['id']);
 
         return redirect('/home');
+    }
+
+    // Returns a card list of only cards that were selected by user
+    public function filterDeck ($cards) {
+        return array_filter(json_decode($cards), function ($card)  {
+            return ($card->selected == true);
+        });
+    }
+
+    //Create Card_deck entries for each card
+    public function createCardDecks ($cards,$deck_id) {
+        foreach ($cards as $card) {
+            $entry = Card_deck::create([
+                'deck_id' => $deck_id,
+                'card_id' => $card->id,
+                'quantity' => $card->quantity
+            ]);
+            
+        };
     }
 
     /**
@@ -173,70 +181,46 @@ class DeckController extends Controller
      */
     public function update(Request $request, Deck $deck)
     {
-        $cards = json_decode($request->userDeck);
+        // Cards contained in the updated deck
+        $cards = $this->filterDeck($request->userDeck);
         
+        // Deck Entry from the previously saved deck
         $deck = Deck::find($request->deck_id);
+        
+        // Card_deck entries from the previously saved deck
         $oldCards = Card_deck::where('deck_id', $request->deck_id)->get();
+        
+        // Check if each of the saved cards is in the updated deck
         foreach ($oldCards as $card) {
             $found = false;
-
+            // Iterate through Updated cards to be saved
             foreach($cards as $key => $value) {
                 if ($value->id == $card->card_id) {
                     $found = true;
                     $currentEntry = (Card_deck::where('deck_id',$request->deck_id)->where('card_id',$value->id)->get())[0];
+                    // Only update the entry if the card quantity has changed
                     if ($currentEntry['quantity'] != $cards[$key]->quantity){
                         $currentEntry['quantity'] = $cards[$key]->quantity;
                     }
                     $currentEntry->save();
-                    
                     break;
                 }            
             }
+
+            // Remove found card from the array to speed up future iterations
             if ($found) {
                 unset($cards[$key]);
             }else {
-                $toDelete = (Card_deck::where('deck_id',$request->deck_id)->where('card_id',$card->card_id)->get())[0];
-                $toDelete->delete();
-                dd($toDelete);
+                
+                // If the card is not a leader card, and was not found, delete the Card_deck entry
+                if ($card->card->type_id > 1){
+                    $toDelete = (Card_deck::where('deck_id',$request->deck_id)->where('card_id',$card->card_id)->get())[0];
+                    $toDelete->delete();
+                };
             }
         }
-        // dd($cards[$key]->name);
-        // // $key = array_search(41, $);
-        // dd($key);
-        // $deck['user_id'] = $request->user_id;
-        // $deck['name'] = $request->name;
-        // $deck['isMonarch'] = $request->isMonarch;
-        // // $deck['lead_id'] = $request->lead_id;
-        // $deck['id'] = $request->deck_id;
-        // $deck->save();
-        
-        ///////////////////
-        // $data = $request->all();
-        
-        // $cards = json_decode($data['userDeck'],true);
-        // $deck = Deck::create([
-        //         'user_id' => $data['user_id'],
-        //         'name' => $data['name'],
-        //         'isMonarch' => $data['isMonarch'],
-        //     ]);
-        
-
-        // $entry = Card_deck::create([
-        //     'deck_id' => $deck['id'],
-        //     'card_id' => $data['lead_id'],
-        //     'quantity' => 1
-        //     ]);
-        // foreach ($cards as $card) {
-        //     if ($card['selected']) {
-        //     $entry = Card_deck::create([
-        //         'deck_id' => $deck['id'],
-        //         'card_id' => $card['id'],
-        //         'quantity' => $card['quantity']
-        //         ]);
-        // }
-        // };
-
-        ///////////////////
+        // Create Card_deck entries for cards added to the deck;
+        $this->createCardDecks($cards,$deck['id']);
         return redirect('/home');
